requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
filelock==3.11.0
frozenlist==1.3.3
googlesearch-python==1.2.3
greenlet==2.0.2
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
langchain==0.0.133
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
Pillow==9.3.0
pydantic==1.10.7
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
soupsieve==2.4
SQLAlchemy==1.4.47
sympy==1.11.1
tenacity==8.2.2
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wincertstore==0.2
yarl==1.8.2
